legacyDemodTop V389 trellisTopFpga2 V389
Added UQPSK and AQPSK modes to the list of supported demod modes to fix a problem with the NASA forward link. The I/Q imbalance was so large (10:1) the constellation locked up with a 45 degree slant. To fix this, we now lock the carrier as if it were a BPSK signal by using a 180 degree modulus phase error when in UQPSK or AUQPSK modes. All other QPSK modes still use the 90 degree modulus phase error.