viterbiTopFpga2 and legacyDemodTop V431:
Viterbi decoding is working in BPSK, QPSK, and OQPSK. 

The option to deskew either I or Q was added to bitsync.v for OQPSK modes.

A bootClk was created as a divide-by-2 from the system clock to make the multiboot process more reliable.

The iTrellis and qTrellis outputs from the legacy bitsync were changed to always select iMF and qMF.

The creation of a sym2xEn was changed in trellis.v to support changes to decoder.v