bs1000Demod V484 and tripleDemod V484:
Added CMA equalizer

Fixed a bug in the digitalPLL. The loop error needed inverting.

Added an `include "addressMap.v" to a couple of multih files to get them to compile correclty for both the tripleDemod and semcoDemod. 

Added a bunch of simulation files

Edited uart.v to get rid of a synthesis warning.