tripleLdpc V621:
Added support for routing the LDPC output through the single rail PCM decoder so it can be used to derandomize the data.

tripleDemod:

Fixed problems with the constellation and eye diagrams. I believe these changes were used to deliver the ULA contract but the changes were never committed. I'm not sure what the version number was in the tripleDemod image as it was delivered.