tripleDemodTop V740:
SOQPSK Trellis works in this version using 1 sample per bit. It also uses a 2-tap matched filter.

As committed, this version supports two versions of DQM. One makes the MSE calculations independent for each demod. The other creates a chip-to-chip interface to pass MSE values from the channel demods to the combiner demod. In this version, the combiner demod calculates it's MSE value from the two channel MSEs. To enable the chip-to-chip version, uncomment the DQM_USE_CHIP_TO_CHIP #define in addressMap.v and use the tripleDemodDqmC2C pinout constraint file.

The Reed Solomon Decoder is included in this build as well but I don't have to ability to test it in Lake Mary.