This commit is mainly about building K325 versions of the images and adding an idCode readback capability so the MCU code can detect which FPGA type is installed.
A bug in the DQM framing is fixed.

All the project files for building separate rate versions of the LDPC decoder are also committed here.

There are also a few works in progress: 1) Adding a compile-time option to lock the carrier loop for SOQPSK on the resampler samples instead of the DDC samples, 2) Adding acqTrack control to the bitsync loop to improve multih acquisition. 3) Adding all the phase error terms to the multiH output to improve acquisition